+++ DOCS/CHANGELOG.md
@@ -12,6 +12,8 @@ The format is based on [Keep a Changelog], and this project adheres to
 - new features: `_char7`, `_char8`, `_char16`, `_char24`, `_char32`.
 - new `Float` and `ExtFloat` method `eval_poly` to evaluate polynomials.
 - new struct `Env` that namespaces `std::env` functions and constants.
+- reexport `core::cell` types.
+- new `CellOption` trait.
 - new traits: `MemPod`.
 
 ### Removed